Revert "Merge branch 'main' into noPCH-5.7"
This reverts commit a3813e13c588a21ef8314e7c9d571500b1ad5e48, reversing
changes made to 93b7e3b832085e4eaa263ee6c04f5e3b210e728d.

+++ Source/PCGExElementsTopology/Public/Elements/PCGExDynamicMeshToClusters.h
@@ -71,7 +71,7 @@ public:
 	 * The value is clamped to be no less than a small positive value to prevent division by zero errors.
 	 */
 	UPROPERTY(BlueprintReadWrite, EditAnywhere, Category = Settings, meta = (PCG_Overridable))
-	float VertexMergeHashTolerance = static_cast<float>(PCGExMesh::DefaultVertexMergeHashTolerance);
+	float VertexMergeHashTolerance = PCGExMesh::DefaultVertexMergeHashTolerance;
 
 	/**
 	 * Use two overlapping spatial hashes to detect vertex proximity. True (default) is more accurate but
